bifrost fix (#153)

+++ packages/executor/src/services/BundlingService/relayers/classic.ts
@@ -115,16 +115,19 @@ export class ClassicRelayer extends BaseRelayer {
       // geth-dev's jsonRpcSigner doesn't support signTransaction
       if (!this.config.testingMode) {
         // check for execution revert
-        try {
-          // eslint-disable-next-line @typescript-eslint/no-unused-vars
-          const { gasLimit, ...txWithoutGasLimit } = transactionRequest;
-          // some chains, like Bifrost, don't allow setting gasLimit in estimateGas
-          await relayer.estimateGas(txWithoutGasLimit);
-        } catch (err) {
-          this.logger.error(err);
-          await this.mempoolService.removeAll(entries);
-          this.reportFailedBundle();
-          return;
+
+        if (!this.networkConfig.skipBundleValidation) {
+          try {
+            // eslint-disable-next-line @typescript-eslint/no-unused-vars
+            const { gasLimit, ...txWithoutGasLimit } = transactionRequest;
+            // some chains, like Bifrost, don't allow setting gasLimit in estimateGas
+            await relayer.estimateGas(txWithoutGasLimit);
+          } catch (err) {
+            this.logger.error(err);
+            await this.mempoolService.removeAll(entries);
+            this.reportFailedBundle();
+            return;
+          }
         }
 
         this.logger.debug(
